+++ xknx/core/group_address_dpt.py
@@ -71,12 +71,12 @@ class GroupAddressDPT:
         try:
             value = transcoder.from_knx(telegram.payload.value)
         except (CouldNotParseTelegram, ConversionError) as err:
-            _GA_DPT_LOGGER.warning(
+            _GA_DPT_LOGGER.debug(
                 "DPT decoding error. Telegram from %s to %s with payload %s can't be decoded by %s: %s",
                 telegram.source_address,
                 telegram.destination_address,
                 telegram.payload.value,
-                transcoder.__name__,
+                transcoder.dpt_name(),
                 err,
             )
             return