+++ xknx/dpt/dpt_17.py
@@ -29,7 +29,7 @@ class DPTSceneNumber(DPTValue1ByteUnsigned):
 
         if not cls._test_boundaries(value):
             raise ConversionError(
-                f"Could not parse {cls.__name__}", value=value, payload=payload
+                f"Could not parse {cls.dpt_name()}", value=value, payload=payload
             )
 
         return value
@@ -44,5 +44,5 @@ class DPTSceneNumber(DPTValue1ByteUnsigned):
             return DPTArray(knx_value)
         except ValueError as err:
             raise ConversionError(
-                f"Could not serialize {cls.__name__}", value=value
+                f"Could not serialize {cls.dpt_name()}", value=value
             ) from err